split batman tests
